use viewmodel to handle functionalities
